refactor(router): move directives into their own folder
