+++ modules/angular2/src/router/directives/router_link.ts
@@ -1,9 +1,9 @@
 import {Directive} from 'angular2/core';
 import {isString} from 'angular2/src/facade/lang';
 
-import {Router} from './router';
-import {Location} from './location/location';
-import {Instruction} from './instruction';
+import {Router} from '../router';
+import {Location} from '../location/location';
+import {Instruction} from '../instruction';
 
 /**
  * The RouterLink directive lets you link to specific parts of your app.
