+++ modules/angular2/src/router/directives/router_link_transform.ts
