+++ modules/angular2/src/router/directives/router_outlet.ts
@@ -14,11 +14,11 @@ import {
   Dependency
 } from 'angular2/core';
 
-import * as routerMod from './router';
-import {ComponentInstruction, RouteParams, RouteData} from './instruction';
-import * as hookMod from './lifecycle/lifecycle_annotations';
-import {hasLifecycleHook} from './lifecycle/route_lifecycle_reflector';
-import {OnActivate, CanReuse, OnReuse, OnDeactivate, CanDeactivate} from './interfaces';
+import * as routerMod from '../router';
+import {ComponentInstruction, RouteParams, RouteData} from '../instruction';
+import * as hookMod from '../lifecycle/lifecycle_annotations';
+import {hasLifecycleHook} from '../lifecycle/route_lifecycle_reflector';
+import {OnActivate, CanReuse, OnReuse, OnDeactivate, CanDeactivate} from '../interfaces';
 
 let _resolveToTrue = PromiseWrapper.resolve(true);
